+++ lib/ash_tui/views/relationships_tab.ex
@@ -12,7 +12,7 @@ defmodule AshTui.Views.RelationshipsTab do
   alias AshTui.State
   alias AshTui.Theme
   alias ExRatatui.Layout.Rect
-  alias ExRatatui.Widgets.{Block, Table}
+  alias ExRatatui.Widgets.{Block, Scrollbar, Table}
 
   @header ["Name", "Type", "Destination", ""]
   @widths [{:min, 12}, {:length, 18}, {:min, 15}, {:length, 12}]
@@ -57,7 +57,26 @@ defmodule AshTui.Views.RelationshipsTab do
       }
     }
 
-    [{table, rect}]
+    viewport_h = max(rect.height - 3, 1)
+    row_count = length(rows)
+
+    scrollbar_widgets =
+      if row_count > viewport_h and selected != nil do
+        [
+          {%Scrollbar{
+             orientation: :vertical_right,
+             content_length: row_count,
+             position: selected,
+             viewport_content_length: viewport_h,
+             thumb_style: Theme.focused_border_style(),
+             track_style: Theme.unfocused_border_style()
+           }, rect}
+        ]
+      else
+        []
+      end
+
+    [{table, rect}] ++ scrollbar_widgets
   end
 
   defp format_type(:belongs_to), do: "\u{2190} belongs_to"